Redundant modules deleted. Some files moved.
