+++ core-android/src/test/java/com/mustafakocer/core_android/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.mustafakocer.data_common
+package com.mustafakocer.core_android
 
 import org.junit.Test
