+++ core-domain/src/main/java/com/mustafakocer/core_common/exception/AppException.kt
